+++ src/components/Spinner.js
@@ -3,8 +3,8 @@ import React from "react";
 import classnames from "../classnames";
 import Circles from "./Circles";
 
-const Spinner = ({ className }) => (
-  <div className={classnames("chq-spn", className)}>
+const Spinner = ({ className, ...props }) => (
+  <div className={classnames("chq-spn", className)} {...props}>
     <Circles />
   </div>
 );